Revert "feat: release student take over (#280)"
This reverts commit e68f9795c049afe63065b26b358ed2cac690f46b.
